+++ airbyte-cdk/python/unit_tests/sources/streams/http/requests_native_auth/test_requests_native_auth.py
@@ -163,6 +163,30 @@ class TestOauth2Authenticator:
 
         assert ("access_token", 1000) == token
 
+        # Test with expires_in as str
+        mocker.patch.object(resp, "json", return_value={"access_token": "access_token", "expires_in": "2000"})
+        token = oauth.refresh_access_token()
+
+        assert ("access_token", 2000) == token
+
+    @pytest.mark.parametrize("error_code", (429, 500, 502, 504))
+    def test_refresh_access_token_retry(self, error_code, requests_mock):
+        oauth = Oauth2Authenticator(
+            f"https://{TestOauth2Authenticator.refresh_endpoint}",
+            TestOauth2Authenticator.client_id,
+            TestOauth2Authenticator.client_secret,
+            TestOauth2Authenticator.refresh_token
+        )
+        requests_mock.post(
+            f"https://{TestOauth2Authenticator.refresh_endpoint}",
+            [
+                {"status_code": error_code}, {"status_code": error_code}, {"json": {"access_token": "token", "expires_in": 10}}
+            ]
+        )
+        token, expires_in = oauth.refresh_access_token()
+        assert (token, expires_in) == ("token", 10)
+        assert requests_mock.call_count == 3
+
     def test_auth_call_method(self, mocker):
         oauth = Oauth2Authenticator(
             token_refresh_endpoint=TestOauth2Authenticator.refresh_endpoint,
@@ -242,6 +266,7 @@ class TestSingleUseRefreshTokenOauth2Authenticator:
             connector_config,
             token_refresh_endpoint="foobar",
         )
+
         authenticator._get_refresh_access_token_response = mocker.Mock(
             return_value={
                 authenticator.get_access_token_name(): "new_access_token",
@@ -250,6 +275,16 @@ class TestSingleUseRefreshTokenOauth2Authenticator:
             }
         )
         assert authenticator.refresh_access_token() == ("new_access_token", 42, "new_refresh_token")
+
+        # Test with expires_in as str
+        authenticator._get_refresh_access_token_response = mocker.Mock(
+            return_value={
+                authenticator.get_access_token_name(): "new_access_token",
+                authenticator.get_expires_in_name(): "1000",
+                authenticator.get_refresh_token_name(): "new_refresh_token",
+            }
+        )
+        assert authenticator.refresh_access_token() == ("new_access_token", 1000, "new_refresh_token")
 
 
 def mock_request(method, url, data):