+++ airbyte-cdk/python/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 0.30.2
+* Low-code CDK: Override refresh_access_token logic DeclarativeOAuthAuthenticator
+
+## 0.30.1
+Releasing using the new release flow. No change to the CDK per se
+
+## 0.30.0
+OAuth: retry refresh access token requests
+
+## 0.29.3
+Low-Code CDK: duration macro added
+
+## 0.29.2
+support python3.8
+
 ## 0.29.1
 Publishing Docker image for source-declarative-manifest